pyproject.toml
[tool.poetry]
name = "cw5"
version = "0.1.0"
description = ""
authors = ["Victor Volkov <vektorn1212@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
